% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.traildb.R
\name{as.traildb}
\alias{as.traildb}
\title{Create traildb objects}
\usage{
as.traildb(X, ...)
}
\arguments{
\item{X}{An object to be converted to a traildb object.}
}
\description{
Create traildb objects
}

