% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{describe_traildb}
\alias{describe_traildb}
\title{Provide a summary of the traildb object}
\usage{
describe_traildb(file)
}
\arguments{
\item{file}{Path to the traildb object (tarball will be automatically unpacked)}
}
\value{
Summary of the traildb object
}
\description{
\code{describe_traildb} returns a summary of the traildb object
}

