% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_dim_traildb}
\alias{get_dim_traildb}
\title{get_dim_traildb returns the list of dimensions}
\usage{
get_dim_traildb(file)
}
\arguments{
\item{file}{Path to the traildb object.}
}
\value{
Vector of dimensions
}
\description{
\code{get_dim_traildb} returns a vector of the dimensions
}

