\name{RTraildb-package}
\alias{RTraildb-package}
\alias{RTraildb}
\docType{package}
\title{
\packageTitle{RTraildb}
}
\description{
\packageDescription{RTraildb}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RTraildb}
\packageIndices{RTraildb}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RTraildb}

Maintainer: \packageMaintainer{RTraildb}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
