% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traildb-class.R
\docType{class}
\name{traildb-class}
\alias{traildb}
\alias{traildb-class}
\title{S4 traildb objects (tbo)}
\description{
traildb objects
}
\section{Creating traildb objects}{


See \code{\link{as.traildb}}.
}

\section{Functions for manipulating traildb objects}{

}

